Rename other directories.
