+++ extensions/ql-vscode/src/view/model-editor/InProgressDropdown.tsx
