+++ extensions/ql-vscode/src/view/model-editor/KindInput.tsx
