+++ extensions/ql-vscode/src/view/model-editor/LibraryRow.tsx
