+++ extensions/ql-vscode/src/view/model-editor/MethodClassifications.tsx
