+++ extensions/ql-vscode/src/view/model-editor/ModeledMethodsList.tsx
