+++ extensions/ql-vscode/src/view/model-editor/ModelingStatusIndicator.tsx
