+++ extensions/ql-vscode/test/unit-tests/model-editor/auto-model.test.ts
