+++ extensions/ql-vscode/test/unit-tests/model-editor/extension-pack-name.test.ts
