+++ extensions/ql-vscode/test/unit-tests/model-editor/shared/modeled-percentage.spec.ts
