+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/data-extensions-editor-queries.test.ts
